split db.js into db.js and models.js, switched to single-init Dexie object
